refactor(cua-computer): move the shared act-params type to the leaf module
